# Added faster and better hfitted method for ETS model